services/jenkins-run-statistics/requirements.txt
boto3
requests-xml
python-dateutil